pyproject.toml
[project]
name = "search-mcp"
version = "0.1.0"
description = "MCP demo for keyword search with Elasticsearch and LLM query planning"
authors = [
    {name = "Kevin Zhao",email = "kevinhfzhao@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli] (>=1.4.1,<2.0.0)",
    "elasticsearch (>=8.17.2,<9.0.0)",
    "openai (>=1.66.3,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "faker (>=37.0.0,<38.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
